requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
django-widget-tweaks
django-jazzmin